refactor(request): extract approval authorization helpers
